feat(day8): add test for updating employee when status is false

+++ src/main/java/org/example/demo/service/EmployeeService.java
@@ -54,6 +54,9 @@ public class EmployeeService {
 
     }
     public void deleteEmployee(long id) {
+        if (!employeeRepositoryImpl.findById(id).get().getStatus()){
+            throw new EmployeeNotExsitingException();
+        }
         employeeRepositoryImpl.delete(employeeRepositoryImpl.findById(id).get());
     }
     public List<Employee> getEmployeeByPage(int page, int size) {
